[oneshot] Surround _system_groupadd RPM macro with parentheses.  Contribute to JB#38494
Expansion of a macro with arguments consumes the rest of the line
including the final newline character. As a result the line following
macro invocation is joined with the very last line of macro's body. In
certain cases this can result in ill formed output.  One option to
prevent this is to make a trailing newline part of macro definition.
Another option is to wrap macro body with parentheses which works better
here considering the use of || in the macro.